Version 0.4 Arranged the framework in a package Created a run.py file for running the server and initlize file Created a new package named market market   ->Created a init file     ->Initializes the app & db     ->Calls for routes file   ->Created a routes.py file     ->Contains all the app routes functions     ->Calls Item class of model file   ->Created a mdodel.py file     ->Initializes the database
