+++ market/model.py
@@ -1,0 +1,13 @@
+from market import db
+
+class Item(db.Model):
+    id = db.Column(db.Integer(),primary_key = True)
+    name = db.Column(db.String(length = 30),nullable = False, unique = True)
+    price = db.Column(db.Integer(), nullable = False)
+    barcode = db.Column(db.String(length = 12), nullable = False, unique = True)
+    description = db.Column(db.String(length = 1024), nullable = False, unique = True)
+
+    #To get the string representation of the database
+    ##Returns the name of the item
+    def __repr__(self):
+        return(f'Item: {self.name}')
