+++ run.py
@@ -1,0 +1,5 @@
+from market import app
+
+#Checks if our run.py executed directly and not imported
+if __name__ == '__main__':
+	app.run(debug = True)
